+++ packages/abjad-web/src/app/layout.tsx
@@ -5,14 +5,10 @@ import React from 'react'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-	title: 'Abjad web app',
+	title: 'Abjad web',
 }
 
-export default function RootLayout({
-																		 children,
-																	 }: {
-	children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
 	return (
 		<html lang="en">
 			<body className={inter.className}>{children}</body>