+++ packages/abjad-web/CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## [0.2.0] 2024-01-16
 ### Changed
-- style the page
+- style web page
+- remove convert button
+- do convert on changes text and select from dropdown
 
 ## [0.1.0] 2024-01-14
 ### Changed