+++ packages/abjad-web/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.2.0] 2024-01-16
+### Changed
+- style web page
+- remove convert button
+- do convert on changes text and select from dropdown
+
 ## [0.1.0] 2024-01-14
 ### Changed
 - use abjad-convert 0.2.1